third_party/blink/web_tests/fast/events/touch/gesture/gesture-tap-near-iframe-expected.txt

Verify that in the absence of any reason to adjust into an iframe, taps are handle in the frame matching the center point of the tap. bug 401504

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Sending GestureTap on top of iframe
PASS sawClick is false


Sending GestureTap above iframe
Received click on DIV#container
PASS sawClick is true


Sending GestureTap to the left of iframe
Received click on DIV#container
PASS sawClick is true


Sending GestureTap to the right of iframe
Received click on DIV#container
PASS sawClick is true


Sending GestureTap below iframe
Received click on DIV#container
PASS sawClick is true


PASS successfullyParsed is true

TEST COMPLETE
